add line for tokens

+++ shipc/chunk.c
@@ -21,9 +21,9 @@ void write_chunk(Chunk* chunk, uint8_t byte) {
 	chunk->count++;
 }
 
-int add_constant(Chunk* chunk, Value constant) {
+uint8_t add_constant(Chunk* chunk, Value constant) {
 	write_value_array(&chunk->constants, constant);
-	return chunk->constants.count - 1;
+	return (uint8_t)(chunk->constants.count - 1);
 }
 
 void free_chunk(Chunk* chunk) {
